created right drawer with current weather

+++ app/src/main/java/pl/marchuck/majormallstrikesback/model/Sys.kt
@@ -1,8 +1,9 @@
 package pl.marchuck.majormallstrikesback.model
 
+
 /**
  * Created by Łukasz Marczak
 
  * @since 11.02.16
  */
-data class Sys(val message: Double, val country: String, val sunrise: Int, val sunset: Int)
+data class Sys(val message: Double, val country: String, val sunrise: Long, val sunset: Long)
